dipdup_indexer/pyproject.toml
[project]
name = "dipdup-indexer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "web3",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
]
